+++ internal/services/user_common_actions_test.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"errors"
 	"testing"
-	"time"
 
-	"github.com/KalessinD/gophermart/internal/common"
 	"github.com/KalessinD/gophermart/internal/models"
 	"github.com/KalessinD/gophermart/internal/repositories/postgresql/mocks"
 	"github.com/KalessinD/gophermart/internal/services"
-	"github.com/golang-jwt/jwt/v5"
 	"go.uber.org/mock/gomock"
 	"golang.org/x/crypto/bcrypt"
 )
@@ -135,52 +132,6 @@ func TestCommonAction_Login(t *testing.T) {
 		err := svc.Login(ctx, userReq)
 		if !errors.Is(err, models.ErrWrongPassword) {
 			t.Errorf("expected ErrWrongPassword, got %v", err)
-		}
-	})
-}
-
-// тестирует генерацию JWT
-func TestCommonAction_GenerateToken(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	mockDB := mocks.NewMockSQLStorageInterface(ctrl)
-	svc := services.NewCommonAction(mockDB)
-
-	t.Run("token generation", func(t *testing.T) {
-		user := &models.User{
-			ID:    "123",
-			Login: "testuser",
-		}
-		expireAt := time.Now().Add(time.Hour)
-		secret := "test-secret"
-
-		tokenStr, err := svc.GenerateToken(user, secret, expireAt)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		// Парсим токен для проверки содержимого
-		claims := &common.Claims{}
-		token, err := jwt.ParseWithClaims(tokenStr, claims, func(token *jwt.Token) (interface{}, error) {
-			if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-				return nil, errors.New("unexpected signing method")
-			}
-			return []byte(secret), nil
-		})
-		if err != nil {
-			t.Fatalf("failed to parse token: %v", err)
-		}
-
-		if !token.Valid {
-			t.Error("token is not valid")
-		}
-
-		if claims.UserID != "123" {
-			t.Errorf("expected userID 123, got %s", claims.UserID)
-		}
-		if claims.Login != "testuser" {
-			t.Errorf("expected login testuser, got %s", claims.Login)
 		}
 	})
 }
